DCMAW-8264: Update session recovery service naming
