+++ src/functions/asyncCredential/sessionService/dynamoDbClient.ts
