+++ src/functions/asyncCredential/sessionService/sessionService.test.ts
@@ -1,0 +1,133 @@
+import { DynamoDBClient, QueryCommand } from "@aws-sdk/client-dynamodb";
+import { SessionService } from "./sessionService";
+import { mockClient } from "aws-sdk-client-mock";
+
+describe("Session Service", () => {
+  let service: SessionService;
+  beforeEach(() => {
+    service = new SessionService("mockTableName", "mockIndexName");
+  });
+
+  describe("Session recovery", () => {
+    describe("Given there is an unexpected error when calling Dynamo DB", () => {
+      it("Returns error response", async () => {
+        const dbMock = mockClient(DynamoDBClient);
+        dbMock.on(QueryCommand).rejects("Mock DB Error");
+
+        const result = await service.getAuthSessionBySub(
+          "mockSub",
+          "mockValidState",
+          3600,
+        );
+
+        expect(result.isError).toBe(true);
+        expect(result.value).toEqual(
+          "Unexpected error when querying session table whilst checking for recoverable session",
+        );
+      });
+    });
+
+    describe("Given Items array is missing", () => {
+      it("Returns success response will value of null", async () => {
+        const dbMock = mockClient(DynamoDBClient);
+        dbMock.on(QueryCommand).resolves({});
+
+        const result = await service.getAuthSessionBySub(
+          "mockSub",
+          "mockValidState",
+          3600,
+        );
+
+        expect(result.isError).toBe(false);
+        expect(result.value).toEqual(null);
+      });
+    });
+
+    describe("Given Items array is empty", () => {
+      it("Returns success response will value of null", async () => {
+        const dbMock = mockClient(DynamoDBClient);
+        dbMock.on(QueryCommand).resolves({ Items: [] });
+
+        const result = await service.getAuthSessionBySub(
+          "mockSub",
+          "mockValidState",
+          3600,
+        );
+
+        expect(result.isError).toBe(false);
+        expect(result.value).toEqual(null);
+      });
+    });
+
+    describe("Given Items array is missing sessionId", () => {
+      it("Returns success response will value of null", async () => {
+        const dbMock = mockClient(DynamoDBClient);
+        dbMock.on(QueryCommand).resolves({
+          Items: [
+            {
+              sub: { S: "mockSub" },
+              state: { S: "mockValidState" },
+            },
+          ],
+        });
+
+        const result = await service.getAuthSessionBySub(
+          "mockSub",
+          "mockValidState",
+          3600,
+        );
+
+        expect(result.isError).toBe(false);
+        expect(result.value).toEqual(null);
+      });
+    });
+
+    describe("Given sessionId value in Items array is empty", () => {
+      it("Returns success response will value of null", async () => {
+        const dbMock = mockClient(DynamoDBClient);
+        dbMock.on(QueryCommand).resolves({
+          Items: [
+            {
+              sub: { S: "mockSub" },
+              state: { S: "mockValidState" },
+              sessionId: { S: "" },
+            },
+          ],
+        });
+
+        const result = await service.getAuthSessionBySub(
+          "mockSub",
+          "mockValidState",
+          3600,
+        );
+
+        expect(result.isError).toBe(false);
+        expect(result.value).toEqual(null);
+      });
+    });
+
+    describe("Given a valid recoverable session is found", () => {
+      it("Returns a success response with sessionId as value", async () => {
+        const dbMock = mockClient(DynamoDBClient);
+        dbMock.on(QueryCommand).resolves({
+          Items: [
+            {
+              sub: { S: "mockSub" },
+              state: { S: "mockValidState" },
+              sessionId: { S: "mockSessionId" },
+            },
+          ],
+        });
+
+        const result = await service.getAuthSessionBySub(
+          "mockSub",
+          "mockValidState",
+          3600,
+        );
+
+        expect(result.isError).toBe(false);
+        expect(result.value).toEqual("mockSessionId");
+      });
+    });
+  });
+});
